+++ crnsynth2/process/dp_stats.py
@@ -1,11 +1,25 @@
 """Compute DP stats from real data."""
-from diffprivlib.tools.quantiles import median
+from typing import Iterable, Union
+
+import pandas as pd
+from diffprivlib.tools.quantiles import median, quantile
 from diffprivlib.tools.utils import mean, std, var
 
 
 def dp_mean(array, epsilon, bounds, random_state=None):
     """Wrapper of mean function from diffprivlib. Computes the DP mean of an array."""
     return mean(array=array, epsilon=epsilon, bounds=bounds, random_state=random_state)
+
+
+def dp_quantile(array, epsilon, bounds, quant, random_state=None):
+    """Wrapper of quantile function from diffprivlib. Computes the DP quantile of an array."""
+    return quantile(
+        array=array,
+        epsilon=epsilon,
+        quant=quant,
+        bounds=bounds,
+        random_state=random_state,
+    )
 
 
 def dp_median(array, epsilon, bounds, random_state=None):
@@ -31,3 +45,60 @@ DP_PARAM_FUNC = {
     "var": dp_var,
     "std": dp_std,
 }
+
+
+class DPParam:
+    """Compute a differential privacy parameter from real data."""
+
+    def __init__(
+        self,
+        stat_name: str,
+        epsilon: float,
+        column: str,
+        bounds: Iterable,
+        random_state: Union[int, None] = None,
+        verbose: int = 1,
+    ):
+        """Initialize the DPParam object.
+
+        Args:
+            stat_name (str): The name of the statistic to compute.
+            epsilon (float): The epsilon value to use for DP computation.
+            column (str): The column to compute the statistic on.
+            bounds (Iterable): The min and max bounds of the stat.
+            random_state (int or None): The random state to use for DP computation.
+            verbose (int, optional): Verbosity level. Defaults to 1.
+        """
+        self.stat_name = stat_name
+        self.epsilon = epsilon
+        self.bounds = bounds
+        self.column = column
+        self.random_state = random_state
+        self.verbose = verbose
+
+    def compute(self, data_real: pd.DataFrame):
+        """Compute a differential privacy parameter from real data.
+
+        Args:
+            data_real (pd.DataFrame): The real data to compute the parameter from.
+
+        Returns:
+            float: The computed parameter.
+        """
+        if self.verbose:
+            print(f"Computing DP parameter {self.stat_name} on column {self.column}")
+
+        # compute dp param
+        dp_param = DP_PARAM_FUNC[self.stat_name](
+            array=data_real[self.column],
+            epsilon=self.epsilon,
+            bounds=self.bounds,
+            random_state=self.random_state,
+        )
+
+        # store param in object to re-use in case you don't want to re-compute
+        self.param_ = dp_param
+        return dp_param
+
+    def __repr__(self):
+        return f"DPParam(stat_name={self.stat_name}, epsilon={self.epsilon}, column={self.column})"
